feat(reflect)!: Cleanup exports
BREAKING CHANGE!

This updates the exports for

- `@rocicorp/reflect`
- `@rocicorp/reflect/client`
- `@rocicorp/reflect/server`

Fixes #362
