--- packages/reflect-client/src/replicache-mod.ts
@@ -1,87 +1,0 @@
-// TODO(arv): Only export the types that are actually used.
-// https://github.com/rocicorp/mono/issues/362
-
-export {
-  ExperimentalMemKVStore,
-  IDBNotFoundError,
-  PullError,
-  TEST_LICENSE_KEY,
-  TransactionClosedError,
-  consoleLogSink,
-  deleteAllReplicacheData,
-  dropAllDatabases,
-  dropDatabase,
-  filterAsyncIterable,
-  getDefaultPuller,
-  isScanIndexOptions,
-  makeIDBName,
-  makeScanResult,
-  mergeAsyncIterables,
-  version,
-} from 'replicache';
-export type {
-  AsyncIterableIteratorToArray,
-  ClientStateNotFoundResponse,
-  Cookie,
-  CreateIndexDefinition,
-  ExperimentalCreateKVStore,
-  ExperimentalDiff,
-  ExperimentalDiffOperation,
-  ExperimentalDiffOperationAdd,
-  ExperimentalDiffOperationChange,
-  ExperimentalDiffOperationDel,
-  ExperimentalIndexDiff,
-  ExperimentalKVRead,
-  ExperimentalKVStore,
-  ExperimentalKVWrite,
-  ExperimentalNoIndexDiff,
-  ExperimentalWatchCallbackForOptions,
-  ExperimentalWatchIndexCallback,
-  ExperimentalWatchIndexOptions,
-  ExperimentalWatchNoIndexCallback,
-  ExperimentalWatchNoIndexOptions,
-  ExperimentalWatchOptions,
-  GetIndexScanIterator,
-  GetScanIterator,
-  HTTPRequestInfo,
-  IndexDefinition,
-  IndexDefinitions,
-  IndexKey,
-  IterableUnion,
-  KeyTypeForScanOptions,
-  LogLevel,
-  LogSink,
-  MutationV0,
-  MutationV1,
-  PatchOperation,
-  PendingMutation,
-  Poke,
-  PullRequest,
-  PullRequestV0,
-  PullRequestV1,
-  PullResponse,
-  PullResponseOKV0,
-  PullResponseOKV1,
-  PullResponseV0,
-  PullResponseV1,
-  Puller,
-  PullerResult,
-  PullerResultV0,
-  PullerResultV1,
-  PushError,
-  PushRequest,
-  PushRequestV0,
-  PushRequestV1,
-  PushResponse,
-  Pusher,
-  PusherResult,
-  RequestOptions,
-  ScanIndexOptions,
-  ScanNoIndexOptions,
-  ScanOptionIndexedStartKey,
-  ScanOptions,
-  ScanResult,
-  SubscribeOptions,
-  UpdateNeededReason,
-  VersionNotSupportedResponse,
-} from 'replicache';
