+++ packages/reflect-server/src/server/options.ts
@@ -1,6 +1,11 @@
-import type {Context, LogLevel, LogSink} from '@rocicorp/logger';
+import {
+  consoleLogSink,
+  type Context,
+  type LogLevel,
+  type LogSink,
+} from '@rocicorp/logger';
 import type {MutatorDefs} from 'reflect-shared';
-import {consoleLogSink, createWorkerDatadogLogSink} from '../mod.js';
+import {createWorkerDatadogLogSink} from '../mod.js';
 import {isTrueEnvValue} from '../util/env.js';
 import type {ReflectServerOptions} from './reflect.js';
