--- packages/reflect-server/src/util/version.ts
@@ -1,6 +1,0 @@
-import {version as sharedVersion} from 'reflect-shared';
-
-/**
- * @deprecated Use `version` from `@rocicorp/reflect` instead.
- */
-export const version = sharedVersion;
