+++ packages/reflect-shared/src/mod.ts
@@ -1,3 +1,54 @@
-export {version} from './version.js';
-
+export {
+  TeeLogSink,
+  consoleLogSink,
+  type LogLevel,
+  type LogSink,
+} from '@rocicorp/logger';
+export {
+  ExperimentalMemKVStore,
+  filterAsyncIterable,
+  isScanIndexOptions,
+  makeIDBName,
+  makeScanResult,
+  mergeAsyncIterables,
+} from 'replicache';
+export type {
+  AsyncIterableIteratorToArray,
+  ClientGroupID,
+  ClientID,
+  ClientStateNotFoundResponse,
+  Cookie,
+  GetIndexScanIterator,
+  GetScanIterator,
+  HTTPRequestInfo,
+  IterableUnion,
+  JSONObject,
+  JSONValue,
+  MaybePromise,
+  MutationV0,
+  MutationV1,
+  MutatorReturn,
+  PatchOperation,
+  Poke,
+  PullRequest,
+  PullRequestV0,
+  PullRequestV1,
+  PullResponse,
+  PullResponseOKV0,
+  PullResponseOKV1,
+  PullResponseV0,
+  PullResponseV1,
+  PushRequest,
+  PushRequestV0,
+  PushRequestV1,
+  PushResponse,
+  ReadonlyJSONObject,
+  ReadonlyJSONValue,
+  RequestOptions,
+  TransactionEnvironment,
+  TransactionReason,
+  UpdateNeededReason,
+  VersionNotSupportedResponse,
+} from 'replicache';
 export type * from './types.js';
+export {version} from './version.js';
